requirements-dev.txt
pytest==6.2.5
pytest-dependency==0.5.1
flake8==4.0.1
yapf==0.31.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
colorama==0.4.4
Deprecated==1.2.13
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tenacity==8.0.1
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3

